+++ source/projects/pktpy/tests/test_cpp2/build.sh
@@ -1,0 +1,5 @@
+TARGET="demo"
+
+echo "compiling ${TARGET} with exception support"
+time clang++ --std=c++17 -fexceptions -I../.. -o "${TARGET}" "${TARGET}.cpp"
+