+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.1.x]
 
+- Added examples of wrapping max api and local functions to `pktpy`.
+
 - Added `anything` method for `pktpy` project.
 
 - Fixed `py` cython build step such that it run only if `api.pyx` is changed.